requirements.txt
requests>=2.31.0
azure-eventhub>=5.13.0
